separate telechat from deepseek
