+++ conversion/__init__.py
@@ -53,7 +53,7 @@ TEXT_MODEL_MAP: dict[str, str] = {
     "DeepseekV2ForCausalLM": "deepseek",
     "DeepseekV3ForCausalLM": "deepseek",
     "DeepseekV32ForCausalLM": "deepseek",
-    "TeleChat4ForCausalLM": "deepseek",
+    "TeleChat4ForCausalLM": "telechat",
     "DFlashDraftModel": "qwen",
     "Qwen3DSparkModel": "qwen",
     "DSparkDraftModel": "qwen",
